@@ -8,31 +8,30 @@
     "test": "jest"
   },
   "dependencies": {
-    "react": "^16.4.1",
-    "react-dom": "^16.4.1"
+    "react": "^16.8.0-alpha.0",
+    "react-dom": "^16.8.0-alpha.0"
   },
   "devDependencies": {
-    "@types/enzyme": "^3.1.12",
-    "@types/enzyme-adapter-react-16": "^1.0.2",
-    "@types/jest": "^23.3.0",
-    "@types/node": "^10.5.2",
-    "@types/prettier": "^1.13.2",
-    "@types/prop-types": "^15.5.3",
-    "@types/react": "^16.4.6",
-    "@types/react-dom": "^16.0.6",
-    "emotion": "9.2.6",
-    "enzyme": "^3.3.0",
-    "enzyme-adapter-react-16": "^1.1.1",
-    "jest": "^23.4.1",
-    "jest-dom": "1.10.0",
-    "parcel-bundler": "^1.9.7",
-    "prettier": "^1.13.7",
-    "react-emotion": "^9.2.6",
-    "serve": "^9.2.0",
-    "ts-jest": "^23.0.1",
-    "tslint-config-airbnb": "5.9.2",
-    "tslint-config-prettier": "^1.13.0",
-    "typescript": "^2.9.2"
+    "@types/enzyme": "^3.1.15",
+    "@types/jest": "^23.3.12",
+    "@types/node": "^10.12.18",
+    "@types/prettier": "^1.15.2",
+    "@types/prop-types": "^15.5.8",
+    "@types/react": "^16.7.18",
+    "@types/react-dom": "^16.0.11",
+    "emotion": "10.0.6",
+    "jest": "^23.6.0",
+    "jest-dom": "3.0.0",
+    "parcel-bundler": "^1.11.0",
+    "prettier": "^1.15.3",
+    "react-emotion": "^10.0.0",
+    "react-testing-library": "^5.4.4",
+    "serve": "^10.1.1",
+    "ts-jest": "^23.10.5",
+    "tslint": "^5.12.1",
+    "tslint-config-airbnb": "5.11.1",
+    "tslint-config-prettier": "^1.17.0",
+    "typescript": "^3.2.2"
   },
   "jest": {
     "moduleFileExtensions": [
@@ -41,11 +40,12 @@
       "js",
       "json"
     ],
+    "testURL": "http://localhost",
     "transform": {
-      ".(ts|tsx)": "<rootDir>/node_modules/ts-jest/preprocessor.js"
+      ".(ts|tsx)": "ts-jest"
     },
     "testMatch": [
-      "**/*.spec.(ts|tsx)"
+      "**/*.test.(ts|tsx)"
     ],
     "testPathIgnorePatterns": [
       "/node_modules/"
